requirements.txt
# Core dependencies
numpy>=1.21.0
jax>=0.4.0
jaxlib>=0.4.0
torch>=2.0.0
scipy>=1.9.0

# Optimization
cvxpy>=1.3.0

# Benchmarking and visualization
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Code generation
google-generativeai>=0.3.0
groq>=0.4.0

# Utilities
requests>=2.28.0
